+++ flake.nix
@@ -153,6 +153,8 @@
               !(nixpkgs.lib.hasInfix "--browser" moduleSystem.config.systemd.services.scorchd.serviceConfig.ExecStart);
             assert nixpkgs.lib.hasInfix "--bind 127.0.0.1:33000"
               moduleSystem.config.systemd.services.scorchd.serviceConfig.ExecStart;
+            assert nixpkgs.lib.hasInfix "--search-engines bing,duckduckgo,naver,wikipedia"
+              moduleSystem.config.systemd.services.scorchd.serviceConfig.ExecStart;
             pkgs.runCommand "scorch-module-check" { } ''
               touch "$out"
             '';